requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pyjwt>=2.4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
bcrypt>=3.2.0
python-dotenv>=0.19.0
cryptography>=3.4.8